app/requirements.txt
APScheduler
pillow
Django>=3.0.7,<4
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework
django-cors-headers
requests
django-after-response
fcm-django<1
tzdata
urllib3==1.25.11
django-sslserver